requirements.txt
# Core
streamlit
streamlit-chat

# LLM
langchain
langchain-groq
langgraph

# Data
python-dotenv
requests
pandas

# Visualization
plotly
kaleido
